+++ mobile/lib/widgets/asset_viewer/info_sheet/exif_image_properties.dart
