+++ mobile/lib/widgets/asset_viewer/info_sheet/exif_location.dart
@@ -1,7 +1,7 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
-import 'package:immich_mobile/widgets/asset_viewer/exif_sheet/exif_map.dart';
+import 'package:immich_mobile/widgets/asset_viewer/info_sheet/exif_map.dart';
 import 'package:immich_mobile/entities/asset.entity.dart';
 import 'package:immich_mobile/entities/exif_info.entity.dart';
