+++ mobile/lib/widgets/asset_viewer/info_sheet/exif_map.dart
