+++ mobile/lib/widgets/asset_viewer/info_sheet/exif_people.dart
